added sweetalert on delete functionility

+++ src/pages/dashboard/ProductRow.jsx
@@ -1,21 +1,46 @@
 import React from 'react'
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 
 export default function ProductRow({ item, onDelete }) {
 
     const { id, name, price, image } = item;
 
-    const handleDelete = async () => {
-        await fetch(`http://localhost:3000/shoe/${id}`, {
-            method: "DELETE"
-        })
-            .then(res => res.json())
-            .then(data => {
-                onDelete(id);
-            })
+    const handleDelete = () => {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:3000/shoe/${id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if (data.length !== 0) {
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                            onDelete(id);
+                        }
+
+                    })
+            }
+        });
+
+
+
     };
-    
+
+
     return (
         <tr>
             <td>1</td>
